requirements.txt
opencv-python
PyTurboJPEG
openvino
scipy
torchvision
ultralytics
autopep8

# mklink python3.8 C:\Users\user\AppData\Local\Programs\Python\Python38\python.exe
# mklink python3.11 C:\Users\user\AppData\Local\Programs\Python\Python311\python.exe
# mklink python3.12 C:\Users\user\AppData\Local\Programs\Python\Python312\python.exe

